renamed ForUpdateExpression to ForUpdateOption
