+++ ktorm-core/src/main/kotlin/org/ktorm/dsl/Query.kt
@@ -767,9 +767,9 @@ public fun Query.joinToString(
  *
  * @since 3.1.0
  */
-public fun Query.forUpdate(forUpdateExpression: ForUpdateExpression?): Query {
+public fun Query.forUpdate(forUpdate: ForUpdateOption?): Query {
     val expr = when (expression) {
-        is SelectExpression -> expression.copy(forUpdate = forUpdateExpression)
+        is SelectExpression -> expression.copy(forUpdate = forUpdate)
         is UnionExpression -> throw IllegalStateException("SELECT FOR UPDATE is not supported in a union expression.")
     }
