+++ ktorm-core/src/main/kotlin/org/ktorm/entity/EntitySequence.kt
@@ -19,7 +19,7 @@ package org.ktorm.entity
 import org.ktorm.database.Database
 import org.ktorm.database.DialectFeatureNotSupportedException
 import org.ktorm.dsl.*
-import org.ktorm.expression.ForUpdateExpression
+import org.ktorm.expression.ForUpdateOption
 import org.ktorm.expression.OrderByExpression
 import org.ktorm.expression.SelectExpression
 import org.ktorm.schema.BaseTable
@@ -1511,6 +1511,6 @@ public fun <E : Any> EntitySequence<E, *>.joinToString(
  * @since 3.1.0
  */
 public fun <E : Any, T : BaseTable<E>> EntitySequence<E, T>.forUpdate(
-    forUpdateExpression: ForUpdateExpression?): EntitySequence<E, T> {
-    return this.withExpression(expression.copy(forUpdate = forUpdateExpression))
+    forUpdate: ForUpdateOption?): EntitySequence<E, T> {
+    return this.withExpression(expression.copy(forUpdate = forUpdate))
 }
