+++ ktorm-core/src/main/kotlin/org/ktorm/expression/SqlExpressions.kt
@@ -19,7 +19,6 @@ package org.ktorm.expression
 import org.ktorm.schema.BooleanSqlType
 import org.ktorm.schema.ColumnDeclaring
 import org.ktorm.schema.SqlType
-import org.ktorm.schema.VarcharSqlType
 
 /**
  * Root class of SQL expressions or statements.
@@ -125,7 +124,7 @@ public data class SelectExpression(
     val groupBy: List<ScalarExpression<*>> = emptyList(),
     val having: ScalarExpression<Boolean>? = null,
     val isDistinct: Boolean = false,
-    val forUpdate: ForUpdateExpression? = null,
+    val forUpdate: ForUpdateOption? = null,
     override val orderBy: List<OrderByExpression> = emptyList(),
     override val offset: Int? = null,
     override val limit: Int? = null,
@@ -134,16 +133,9 @@ public data class SelectExpression(
 ) : QueryExpression()
 
 /**
- * For Update expression, implementations are in the MySql and Postgres Dialects.
+ * ForUpdateOption, implementations are in the MySql and Postgres Dialects.
  */
-public abstract class ForUpdateExpression : ScalarExpression<String>() {
-    override val isLeafNode: Boolean
-        get() = true
-    override val extraProperties: Map<String, Any>
-        get() = emptyMap()
-    override val sqlType: SqlType<String>
-        get() = VarcharSqlType
-}
+public interface ForUpdateOption
 
 /**
  * Union expression, represents a `union` statement of SQL.
