+++ ktorm-support-postgresql/src/test/kotlin/org/ktorm/support/postgresql/PostgreSqlTest.kt
@@ -398,7 +398,7 @@ class PostgreSqlTest : BaseTest() {
             val employee = database
                 .sequenceOf(Employees, withReferences = false)
                 .filter { it.id eq 1 }
-                .forUpdate(PostgresForUpdateExpression.NoWait)
+                .forUpdate(PostgresForUpdateOption.NoWait)
                 .first()
 
             val future = Executors.newSingleThreadExecutor().submit {
